Staging env links to all staged PRs with the PR ID as the label.
