+++ Dockerfile
@@ -39,6 +39,7 @@ COPY --from=build /app/.pnp.cjs /app/.pnp.cjs
 COPY --from=build /app/.pnp.loader.mjs /app/.pnp.loader.mjs
 COPY --from=build /app/package.json /app/package.json
 COPY --from=build /app/yarn.lock /app/yarn.lock
+COPY --from=build /app/runtime-utils /app/runtime-utils
 RUN yarn install
 
 COPY --from=flyio/litefs /usr/local/bin/litefs /app/litefs
