+++ staging-server/index.tsx
@@ -1,0 +1,38 @@
+import type { ErrorRequestHandler } from "express"
+import * as React from "react"
+import express from "express"
+import { renderToString } from "react-dom/server"
+import { getApps } from "./apps"
+import LandingPage from "./components/LandingPage"
+
+const catchAllErrors: ErrorRequestHandler = (err, req, res, next) => {
+  console.error(err.stack)
+  res.status(500).send("Error")
+}
+
+const run = async () => {
+  const app = express()
+
+  app.use("/healthcheck", (req, res) => res.status(200).send("OK"))
+
+  const stagedAppsDir = process.env.APP_STAGING_DIR ?? ""
+  const apps = await getApps(stagedAppsDir)
+  app.use("/", (req, res) => {
+    const html = renderToString(<LandingPage apps={apps} />)
+
+    res.status(200).send(html)
+  })
+
+  app.use(catchAllErrors)
+
+  const port = process.env.PORT ?? 3001
+  app.listen(port, () => {
+    console.log(`Staging express server listening on port ${port}`)
+  })
+}
+
+if (require.main === module) {
+  run()
+}
+
+export default run
