+++ staging-server/types.ts
@@ -1,0 +1,7 @@
+type StagedApp = {
+  appDirectory: string
+  prId: string
+  port: number
+}
+
+export type { StagedApp }
